+++ packages/connectivity_plus/connectivity_plus_linux/test/connectivity_plus_linux_test.dart
@@ -1,22 +1,22 @@
 import 'package:connectivity_plus_linux/src/connectivity_real.dart';
-import 'package:connectivity_plus_linux/src/network_manager.dart';
 import 'package:connectivity_plus_platform_interface/connectivity_plus_platform_interface.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
+import 'package:nm/nm.dart';
 
 import 'connectivity_plus_linux_test.mocks.dart';
 
-@GenerateMocks([], customMocks: [
-  MockSpec<NetworkManager>(returnNullOnMissingStub: true),
-])
+@GenerateMocks([NetworkManagerClient])
 void main() {
   test('wireless', () async {
     final linux = ConnectivityLinux();
-    linux.createManager = () {
-      final manager = MockNetworkManager();
-      when(manager.getType()).thenAnswer((_) => Future.value('wireless'));
-      return manager;
+    linux.createClient = () {
+      final client = MockNetworkManagerClient();
+      when(client.connectivity)
+          .thenReturn(NetworkManagerConnectivityState.full);
+      when(client.primaryConnectionType).thenReturn('wireless');
+      return client;
     };
     expect(
         linux.checkConnectivity(), completion(equals(ConnectivityResult.wifi)));
@@ -24,10 +24,11 @@ void main() {
 
   test('no connectivity', () async {
     final linux = ConnectivityLinux();
-    linux.createManager = () {
-      final manager = MockNetworkManager();
-      when(manager.getType()).thenAnswer((_) => Future.value(''));
-      return manager;
+    linux.createClient = () {
+      final client = MockNetworkManagerClient();
+      when(client.connectivity)
+          .thenReturn(NetworkManagerConnectivityState.none);
+      return client;
     };
     expect(
         linux.checkConnectivity(), completion(equals(ConnectivityResult.none)));
@@ -35,15 +36,17 @@ void main() {
 
   test('connectivity changes', () {
     final linux = ConnectivityLinux();
-    linux.createManager = () {
-      final manager = MockNetworkManager();
-      when(manager.getType()).thenAnswer((_) {
-        return Future.value('wireless');
+    linux.createClient = () {
+      final client = MockNetworkManagerClient();
+      when(client.connectivity)
+          .thenReturn(NetworkManagerConnectivityState.full);
+      when(client.primaryConnectionType).thenReturn('wireless');
+      when(client.propertiesChanged).thenAnswer((_) {
+        when(client.connectivity)
+            .thenReturn(NetworkManagerConnectivityState.none);
+        return Stream.value(['Connectivity']);
       });
-      when(manager.subscribeTypeChanged()).thenAnswer((_) {
-        return Stream.value('');
-      });
-      return manager;
+      return client;
     };
     expect(linux.onConnectivityChanged,
         emitsInOrder([ConnectivityResult.wifi, ConnectivityResult.none]));