+++ packages/network_info_plus/network_info_plus_linux/lib/src/network_info_real.dart
@@ -1,78 +1,125 @@
 import 'dart:async';
 
+import 'package:collection/collection.dart';
 import 'package:network_info_plus_platform_interface/network_info_plus_platform_interface.dart';
 import 'package:meta/meta.dart';
-
-import 'network_manager.dart';
+import 'package:nm/nm.dart';
 
 // Used internally
 // ignore_for_file: public_member_api_docs
 
-typedef _DeviceGetter = Future<String?> Function(NMDevice device);
-typedef _ConnectionGetter = Future<String?> Function(NMConnection connection);
+typedef _DeviceGetter = Future<String?> Function(NetworkManagerDevice? device);
+typedef _ConnectionGetter = Future<String?> Function(
+    NetworkManagerActiveConnection? connection);
 
 @visibleForTesting
-typedef NetworkManagerFactory = NetworkManager Function();
+typedef NetworkManagerClientFactory = NetworkManagerClient Function();
 
 /// The Linux implementation of NetworkInfoPlatform.
 class NetworkInfoLinux extends NetworkInfoPlatform {
   /// Obtains the wifi name (SSID) of the connected network
   @override
   Future<String?> getWifiName() {
-    return _getConnectionValue((connection) => connection.getId());
+    return _getConnectionValue((connection) async => connection?.id);
   }
 
-  /// Obtains the IP address of the connected wifi network
+  /// Obtains the IP v4 address of the connected wifi network
   @override
   Future<String?> getWifiIP() {
-    return _getDeviceValue((device) => device.getIp4());
+    return _getConnectionValue(
+      (connection) async => _getIpAddress(connection?.ip4Config?.addressData),
+    );
+  }
+
+  /// Obtains the IP v6 address of the connected wifi network
+  @override
+  Future<String?> getWifiIPv6() {
+    return _getConnectionValue(
+      (connection) async => _getIpAddress(connection?.ip6Config?.addressData),
+    );
   }
 
   /// Obtains the wifi BSSID of the connected network.
   @override
   Future<String?> getWifiBSSID() {
-    return _getDeviceValue((device) {
-      return device
-          .asWirelessDevice()
-          .then((wireless) => wireless.getHwAddress());
-    });
+    return _getDeviceValue((device) async => device?.wireless?.permHwAddress);
+  }
+
+  /// Obtains the submask of the connected wifi network
+  @override
+  Future<String?> getWifiSubmask() {
+    return _getConnectionValue(
+      (connection) async => _getSubnetMask(connection?.ip4Config?.addressData),
+    );
+  }
+
+  /// Obtains the gateway IP address of the connected wifi network
+  @override
+  Future<String?> getWifiGatewayIP() {
+    return _getConnectionValue(
+      (connection) async => connection?.ip4Config?.gateway,
+    );
+  }
+
+  /// Obtains the broadcast of the connected wifi network
+  @override
+  Future<String?> getWifiBroadcast() {
+    return _getConnectionValue(
+      (connection) async => _getBroadcast(connection?.ip4Config?.addressData),
+    );
   }
 
   Future<String?> _getDeviceValue(_DeviceGetter getter) {
     return _getConnectionValue((connection) {
-      return connection.createDevice().then((device) {
-        return getter(device);
-      });
+      final device = connection?.devices
+          .firstWhereOrNull((device) => device.wireless != null);
+      return getter(device);
     });
   }
 
-  Future<String?> _getConnectionValue(_ConnectionGetter getter) {
-    return _ref().createConnection().then((connection) {
-      return getter(connection);
-    }).whenComplete(_deref);
+  Future<String?> _getConnectionValue(_ConnectionGetter getter) async {
+    final client = createClient();
+    await client.connect();
+    final value = getter(client.primaryConnection);
+    await client.close();
+    return value;
   }
 
-  int _refCount = 0;
-  NetworkManager? _manager;
-
-  NetworkManager _ref() {
-    _manager ??= createManager();
-    ++_refCount;
-    return _manager!;
+  String? _getIpAddress(List<Map<String, dynamic>>? data) {
+    return data?.firstOrNull?['address'] as String;
   }
 
-  void _deref() {
-    // schedules an asynchronous disposal when the last reference is removed
-    if (--_refCount == 0) {
-      scheduleMicrotask(() {
-        if (_refCount == 0) {
-          _manager!.dispose();
-          _manager = null;
-        }
-      });
-    }
+  String? _getSubnetMask(List<Map<String, dynamic>>? data) {
+    final prefix = data?.firstOrNull?['prefix'] as int;
+    final mask = 0xffffffff >> (32 - prefix);
+    return mask.toIpString();
+  }
+
+  String? _getBroadcast(List<Map<String, dynamic>>? data) {
+    final ip = _getIpAddress(data)?.toIpInt() ?? 0;
+    final mask = _getSubnetMask(data)?.toIpInt() ?? 0;
+    return (ip | (mask ^ 0xffffffff)).toIpString();
   }
 
   @visibleForTesting
-  NetworkManagerFactory createManager = () => NetworkManager.system();
+  NetworkManagerClientFactory createClient = () => NetworkManagerClient();
+}
+
+extension _IpInt on int {
+  int byteAt(int i) => (this >> (i * 8)) & 0xff;
+  String toIpString() => '${byteAt(0)}.${byteAt(1)}.${byteAt(2)}.${byteAt(3)}';
+}
+
+extension _IpString on String {
+  int toIpInt() {
+    final parts = split('.');
+    return parts.intAtOrZero(3) << 24 |
+        parts.intAtOrZero(2) << 16 |
+        parts.intAtOrZero(1) << 8 |
+        parts.intAtOrZero(0);
+  }
+}
+
+extension _IntOrZero on List<String> {
+  int intAtOrZero(int i) => i < 0 || i >= length ? 0 : int.parse(this[i]);
 }