+++ packages/android_intent_plus/CHANGELOG.md
@@ -1,3 +1,16 @@
+## 2.1.0
+
+- migrate integration_test to flutter sdk
+
+## 2.0.0
+
+- increase flutter SDK version to 1.20.0
+
+## 1.0.3
+
+- clean up ios folder
+- rename channel name to dev.fluttercommunity.plus in java tests
+
 ## 1.0.2
 
 - Android: migrate to mavenCentral