+++ packages/battery_plus/battery_plus/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.2.0
+
+- migrate integration_test to flutter sdk
+
+## 1.1.1
+
+- Fix: Add break statements for unknown battery state in Android and iOS implementations
+
 ## 1.1.0
 
 - Android, iOS, Windows : add getter for power save mode state