+++ packages/connectivity_plus/connectivity_plus/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 1.2.0
+
+- migrate integration_test to flutter sdk
+
+## 1.1.0
+
+- Add ethernet as connectivity result. Supported on Android, Linux, macOS, and Web
+
+## 1.0.8
+
+- Web: Fix to show `ConnectivityResult.mobile` instead of `ConnectivityResult.wifi` when the network is connected to 4g
+
+## 1.0.7
+
+- Android: Fix to detect no-network when Wi-Fi is disconnected on Android M and above
+
 ## 1.0.6
 
 - Android: Fix unregisterReceiver call for API < N