+++ packages/connectivity_plus/connectivity_plus_web/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.1.0+1
+
+- Fix platform interface dependency
+
+## 1.1.0
+
+- Add ethernet as connectivity result
+
 ## 1.0.2
 
 - Update connectivity plus